requirements.txt
crewai>=0.11.0
openai>=1.12.0
python-dotenv>=1.0.0
llama-index>=0.9.0
# Standard library dependencies (no need to install):
# - datetime
# - json
# - re
# - argparse 